Fix various issues related to compaction

+++ server/commitlog/segment.go
@@ -50,6 +50,8 @@ type Segment struct {
 	Index          *Index
 	BaseOffset     int64
 	nextOffset     int64
+	firstOffset    int64
+	lastOffset     int64
 	firstWriteTime int64
 	lastWriteTime  int64
 	position       int64
@@ -64,11 +66,7 @@ type Segment struct {
 	sync.RWMutex
 }
 
-func NewSegment(path string, baseOffset, maxBytes int64, isNew bool, args ...interface{}) (*Segment, error) {
-	var suffix string
-	if len(args) != 0 {
-		suffix = args[0].(string)
-	}
+func NewSegment(path string, baseOffset, maxBytes int64, isNew bool, suffix string) (*Segment, error) {
 	s := &Segment{
 		maxBytes:   maxBytes,
 		BaseOffset: baseOffset,
@@ -78,11 +76,8 @@ func NewSegment(path string, baseOffset, maxBytes int64, isNew bool, args ...int
 		waiters:    make(map[contextReader]chan struct{}),
 	}
 	// If this is a new segment, ensure the file doesn't already exist.
-	if isNew {
-		if _, err := os.Stat(s.logPath()); err == nil {
-			// Segment file already exists.
-			return nil, ErrSegmentExists
-		}
+	if isNew && exists(s.logPath()) {
+		return nil, ErrSegmentExists
 	}
 	log, err := os.OpenFile(s.logPath(), os.O_RDWR|os.O_CREATE|os.O_APPEND, 0666)
 	if err != nil {
@@ -104,6 +99,7 @@ func NewSegment(path string, baseOffset, maxBytes int64, isNew bool, args ...int
 // Initialization is:
 // - Initialize Index position
 // - Initialize Segment nextOffset
+// - Initialize firstOffset/lastOffset
 // - Initialize firstWriteTime/lastWriteTime
 func (s *Segment) setupIndex() (err error) {
 	s.Index, err = NewIndex(options{
@@ -119,13 +115,15 @@ func (s *Segment) setupIndex() (err error) {
 	}
 	// If lastEntry is nil, the index is empty.
 	if lastEntry != nil {
-		s.nextOffset = lastEntry.Offset + 1
+		s.lastOffset = lastEntry.Offset
+		s.nextOffset = s.lastOffset + 1
 		s.lastWriteTime = lastEntry.Timestamp
-		// Read the first entry to get firstWriteTime.
+		// Read the first entry to get firstOffset and firstWriteTime.
 		var firstEntry Entry
 		if err := s.Index.ReadEntryAtFileOffset(&firstEntry, 0); err != nil {
 			return err
 		}
+		s.firstOffset = firstEntry.Offset
 		s.firstWriteTime = firstEntry.Timestamp
 	}
 	return nil
@@ -170,10 +168,28 @@ func (s *Segment) NextOffset() int64 {
 	return s.nextOffset
 }
 
+func (s *Segment) FirstOffset() int64 {
+	s.RLock()
+	defer s.RUnlock()
+	return s.firstOffset
+}
+
+func (s *Segment) LastOffset() int64 {
+	s.RLock()
+	defer s.RUnlock()
+	return s.lastOffset
+}
+
 func (s *Segment) Position() int64 {
 	s.RLock()
 	defer s.RUnlock()
 	return s.position
+}
+
+func (s *Segment) IsEmpty() bool {
+	s.RLock()
+	defer s.RUnlock()
+	return s.firstWriteTime == 0
 }
 
 func (s *Segment) WriteMessageSet(ms []byte, entries []*Entry) error {
@@ -199,9 +215,13 @@ func (s *Segment) Write(p []byte, entries []*Entry) (n int, err error) {
 	s.nextOffset += int64(numMsgs)
 	s.position += int64(n)
 	if s.firstWriteTime == 0 {
-		s.firstWriteTime = entries[0].Timestamp
+		first := entries[0]
+		s.firstOffset = first.Offset
+		s.firstWriteTime = first.Timestamp
 	}
-	s.lastWriteTime = entries[numMsgs-1].Timestamp
+	last := entries[numMsgs-1]
+	s.lastOffset = last.Offset
+	s.lastWriteTime = last.Timestamp
 	s.notifyWaiters()
 	return n, nil
 }
@@ -244,7 +264,8 @@ func (s *Segment) removeWaiter(r contextReader) {
 	s.Unlock()
 }
 
-// Close a segment such that it can no longer be read from or written to.
+// Close a segment such that it can no longer be read from or written to. This
+// operation is idempotent.
 func (s *Segment) Close() error {
 	s.Lock()
 	defer s.Unlock()
@@ -252,6 +273,9 @@ func (s *Segment) Close() error {
 }
 
 func (s *Segment) close() error {
+	if s.closed {
+		return nil
+	}
 	if err := s.log.Close(); err != nil {
 		return err
 	}
@@ -262,27 +286,32 @@ func (s *Segment) close() error {
 	return nil
 }
 
-// Cleaner creates a cleaner segment for this segment.
-func (s *Segment) Cleaner() (*Segment, error) {
+// Cleaned creates a cleaned segment for this segment.
+func (s *Segment) Cleaned() (*Segment, error) {
 	return NewSegment(s.path, s.BaseOffset, s.maxBytes, true, cleanedSuffix)
 }
 
+// Truncated creates a truncated segment for this segment.
+func (s *Segment) Truncated() (*Segment, error) {
+	return NewSegment(s.path, s.BaseOffset, s.maxBytes, true, truncatedSuffix)
+}
+
 // Replace replaces the given segment with the callee.
-func (s *Segment) Replace(old *Segment) (err error) {
+func (s *Segment) Replace(old *Segment) error {
 	s.Lock()
 	defer s.Unlock()
 	old.Lock()
 	defer old.Unlock()
-	if err = old.close(); err != nil {
+	if err := old.close(); err != nil {
 		return err
 	}
-	if err = s.close(); err != nil {
+	if err := s.close(); err != nil {
 		return err
 	}
-	if err = os.Rename(s.logPath(), old.logPath()); err != nil {
+	if err := os.Rename(s.logPath(), old.logPath()); err != nil {
 		return err
 	}
-	if err = os.Rename(s.indexPath(), old.indexPath()); err != nil {
+	if err := os.Rename(s.indexPath(), old.indexPath()); err != nil {
 		return err
 	}
 	s.suffix = ""
@@ -345,11 +374,15 @@ func (s *Segment) Delete() error {
 	}
 	s.Lock()
 	defer s.Unlock()
-	if err := os.Remove(s.log.Name()); err != nil {
-		return err
+	if exists(s.log.Name()) {
+		if err := os.Remove(s.log.Name()); err != nil {
+			return err
+		}
 	}
-	if err := os.Remove(s.Index.Name()); err != nil {
-		return err
+	if exists(s.Index.Name()) {
+		if err := os.Remove(s.Index.Name()); err != nil {
+			return err
+		}
 	}
 	return nil
 }
